Add select operation and refactoring
Closes #1.
